+++ framework/packages/abstract-client/tests/mock_service.rs
@@ -1,4 +1,5 @@
 use abstract_interface::{RegisteredModule, ServiceDeployer};
+use abstract_std::objects::dependency::StaticDependency;
 use abstract_testing::prelude::TEST_VERSION;
 use cosmwasm_std::{to_json_binary, Empty};
 use cw_orch::{contract::Contract, prelude::*};
@@ -65,6 +66,10 @@ impl<Chain> RegisteredModule for MockService<Chain> {
 
     fn module_version<'a>() -> &'a str {
         TEST_VERSION
+    }
+
+    fn dependencies<'a>() -> &'a [StaticDependency] {
+        &[]
     }
 }